// packages/cosmic-web/src/components/header/select/interface.ts
export type RawValueType = string | number;
// ...
export interface TMenuItem {
    content: RawValueType;
    disabled?: boolean;
    empty?: boolean;
    value?: RawValueType;
    title?: string;
}

export interface DropdownMenuProps {
    context?: BaseContext;
    inputValue?: string;
}

export interface DropdownMenuState {
    activeKey: string;
    context: BaseContext;
    inputValue: string;
}

export interface DropdownComputed {
    menuItems: () => TMenuItem[] | undefined;
    selectedKeys: () => RawValueType[];
}

// Select
type TimerType = number | null;
type SelectSize = Exclude<SizeType, 'middle'>;

interface SelectPrivateProps {
    _focused: boolean;
    _mouseDown: boolean;

    focusTimer: TimerType;
    comboboxTimer: TimerType;
    blurTimer: TimerType;
    inputElement: HTMLInputElement;
    modeConfig: ModeConfig;
    backfillValue: string;
    inputValue: string;
    ariaId: string;
    options?: TOptions;
    optionGroups: TOptionGroups;
    optionsInfo: TOptionsInfo;
    open: boolean;
    value: DefaultValueType;
}
// ...
export interface SelectProps extends BaseSelectProps, Omit<SelectPrivateProps, 'options' | 'open' | 'value'>{
    // dropdown
    dropdownStyle?: {
        [key: string]: string;
    };
    dropdownClassName?: string;
    dropdownRender?: (menu: san.ANode) => san.ANode;
    dropdownMenuStyle?: {
        [key: string]: string;
    };

    // >>> Selector
    maxTagTextLength?: number;
    maxTagCount?: number;
    maxTagPlaceholder?: string | ((omittedValues: RawValueType[]) => string);

    // >>> Open
    open?: boolean;
    defaultOpen?: boolean;

    // >>> Icons
    showArrow?: boolean;

    // >>> Mode
    mode?: Mode;

    notFoundContent?: string;
    placeholder?: string;

    options?: DefaultOptionType[];
    filterOption?: boolean | FilterFun<DefaultOptionType>;
    optionLabelProp?: string;

    value?: DefaultValueType ;
    defaultValue?: DefaultValueType ;
    firstActiveValue?: DefaultValueType;
}
// ...
export type SelInstanceProps = SelectState & SelectProps & SelectPrivateProps & {
    [key: string]: any;
};

export type BaseContext = Partial<SelInstanceProps>;

export interface SelectState extends Required<BaseSelectProps> {
    context: BaseContext;
    dropdownPrefixCls: string;
    dropdownWidth: number;
    options: TOptions;
    optionGroups: TOptionGroups;
    optionsInfo: TOptionsInfo;
    showAction: string[];
}